# docs/swagger.yaml
basePath: /api
definitions:
  ent.User:
    properties:
      createTime:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      password:
        description: Password holds the value of the "password" field.
        type: string
      updateTime:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
      username:
        description: Username holds the value of the "username" field.
        type: string
    type: object
  presenter.ErrorResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  presenter.SaveUserReq:
    properties:
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
  presenter.UserLoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  presenter.UserLoginResp:
    properties:
      accessToken:
        type: string
    type: object
host: 127.0.0.1:8000
info:
  contact: {}
  description: This is an API
  title: FiberTemplate
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: User login with username and password returns an access token
      parameters:
      - description: LoginInfo
        in: body
        name: loginReq
        required: true
        schema:
          $ref: '#/definitions/presenter.UserLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.UserLoginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.ErrorResponse'
      summary: user login
      tags:
      - Auth
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Get user details by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/presenter.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.ErrorResponse'
      summary: Get user by ID
      tags:
      - User
  /users:
    post:
      consumes:
      - application/json
      description: Determine whether the user is a new or updated user based on the
        ID in the request body
      parameters:
      - description: UserInfo
        in: body
        name: saveUserReq
        required: true
        schema:
          $ref: '#/definitions/presenter.SaveUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/presenter.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.ErrorResponse'
      summary: Create or update users
      tags:
      - User
swagger: "2.0"